+++ lib/src/features/dashboard/logic/navigation_bar_item.dart
