+++ lib/src/router/router_name.dart
@@ -1,5 +1,4 @@
 enum AppRouteNames {
-  // home(path: '/'),
   login(path: '/login'),
   signUp(path: '/signUp'),
   survey(path: '/survey'),
